End Screen
I added the last war and the mechanic for comparing the 2 cards. The player with the lower card will lose a life. Once a player gets to 0 lives the end screen will pop up and announce the winner.
I still want to add a shop where you can buy cards.